Move ex-transaction classes to service package
