+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/runtime/ServiceRuntimeAdapter.java
@@ -23,8 +23,8 @@ import com.exonum.binding.common.runtime.ServiceArtifactId;
 import com.exonum.binding.core.blockchain.BlockchainData;
 import com.exonum.binding.core.proxy.Cleaner;
 import com.exonum.binding.core.proxy.CloseFailuresException;
+import com.exonum.binding.core.service.ExecutionException;
 import com.exonum.binding.core.storage.database.Snapshot;
-import com.exonum.binding.core.transaction.ExecutionException;
 import com.exonum.messages.core.runtime.Base.ArtifactId;
 import com.exonum.messages.core.runtime.Base.InstanceSpec;
 import com.exonum.messages.core.runtime.Lifecycle.InstanceStatus;
