+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/service/ExecutionPreconditions.java
@@ -5,7 +5,7 @@
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *     https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.exonum.binding.core.transaction;
+package com.exonum.binding.core.service;
 
 import static com.google.common.base.Strings.lenientFormat;
