+++ exonum-java-binding/core/src/main/java/com/exonum/binding/core/service/Schema.java
@@ -17,7 +17,6 @@
 package com.exonum.binding.core.service;
 
 import com.exonum.binding.core.blockchain.Block;
-import com.exonum.binding.core.transaction.ExecutionContext;
 import com.exonum.binding.core.transaction.Transaction;
 
 /**
