+++ exonum-java-binding/core/src/test/java/com/exonum/binding/core/runtime/ServiceRuntimeIntegrationTest.java
@@ -43,11 +43,11 @@ import com.exonum.binding.core.proxy.Cleaner;
 import com.exonum.binding.core.proxy.CloseFailuresException;
 import com.exonum.binding.core.service.BlockCommittedEvent;
 import com.exonum.binding.core.service.Configuration;
+import com.exonum.binding.core.service.ExecutionContext;
 import com.exonum.binding.core.storage.database.Database;
 import com.exonum.binding.core.storage.database.Fork;
 import com.exonum.binding.core.storage.database.Snapshot;
 import com.exonum.binding.core.storage.database.TemporaryDb;
-import com.exonum.binding.core.transaction.ExecutionContext;
 import com.exonum.messages.core.runtime.Lifecycle.InstanceMigration;
 import com.exonum.messages.core.runtime.Lifecycle.InstanceStatus;
 import com.exonum.messages.core.runtime.Lifecycle.InstanceStatus.Simple;
