+++ exonum-java-binding/qa-service/src/main/java/com/exonum/binding/qaservice/QaService.java
@@ -19,10 +19,10 @@ package com.exonum.binding.qaservice;
 import com.exonum.binding.common.crypto.PublicKey;
 import com.exonum.binding.common.hash.HashCode;
 import com.exonum.binding.core.service.Configurable;
+import com.exonum.binding.core.service.ExecutionContext;
+import com.exonum.binding.core.service.ExecutionException;
 import com.exonum.binding.core.service.Node;
 import com.exonum.binding.core.service.Service;
-import com.exonum.binding.core.transaction.ExecutionContext;
-import com.exonum.binding.core.transaction.ExecutionException;
 import com.exonum.binding.core.transaction.RawTransaction;
 import com.exonum.binding.qaservice.transactions.TxMessageProtos;
 import com.exonum.messages.core.Blockchain.Config;
